backend/ai/utils/preprocess.py
from PIL import Image
import torch
from torchvision import transforms

def build_transform(img_size=224):
    return transforms.Compose([
        transforms.Resize((img_size, img_size)),
        transforms.ToTensor(),
        transforms.Normalize(mean=[0.485,0.456,0.406], std=[0.229,0.224,0.225])
    ])

def load_image(file):
    img = Image.open(file).convert("RGB")
    return img
