backend/ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pillow==10.4.0
torch==2.3.1
torchvision==0.18.1
numpy==1.26.4
pydantic==2.7.4
